:sparkles: #27 logging用のpackageを分離、全てのファイルは `src` に移動
